fix: update dependencies for peerDependency warnings
peerDependencies in the packages were throwing warnings, potentially related to
https://github.com/yarnpkg/yarn/issues/5810. In this commit, we rework the peerDependency
definitions to align better with the conventions outlined in the main peerDependency documentation,
and we define devDependencies both at the top level and at the package level, since package-level is
technically closer to "correct" but it doesn't satisfy the checks in yarn yet. In addition, we've
bumped a few versions and aligned across the packages where duplication occurs, which led to our
tests running into https://github.com/yannickcr/eslint-plugin-react/issues/2329 for our "detect"
mode in eslint-plugin-react.