improvement(eslint-config): only set env.jest to true for test files
Previously eslint-config set the jest env to true for all files. This meant that eslint incorrectly
allowed things like references to the jest global. Now, the package uses an override to only apply
this environment for files that look like test files. The pattern used matches the default value of
testMatch in Jest, which is the configuration parameter used to determine which files are matched as
test suites. The only addition to this pattern is matching anything in a __mocks__ directory, which
is the directory name Jest uses to identify mocks (this name cannot be overridden).

cf. https://jestjs.io/docs/en/configuration#testmatch-arraystring

BREAKING CHANGE: This is a breaking change for anyone who has overwritten the value of testMatch in
their Jest configuration, as the override in the eslint config assumes that the default pattern is
used for matching test files. However, if this configuration has not been overwritten, the only
effect of this change will be to expose new linting errors if Jest globals are being incorrectly
referenced in non test/mock code.