fix(web-scripts): positional args were being stripped out
positional args were not making there way through to the web-scripts commands, because commander 5
changed how we need to handle args. there is now conveniently an `args` array on the command which
contains all args (positional or flagged) in order, which seems to be more reliably correct from
what the user inputs compared to using the parseOptions(process.argv) method.