feat(eslint-config): remove jest-specific configuration
This commit removes all the Jest-specific configuration from @spotify/eslint-config in favour of
adding a jest-specific package. This means that it will be possible to use a combination of
@spotify/eslint-config-* packages to allow Jest globals in test code but disallow them in non-test
code (something originally discussed in #30). It also makes it easier to have other test
environments like Cypress coexisting with Jest in a single repository.

BREAKING CHANGE: This change will cause the previously-documented default eslint config, and any
others that rely on Jest globals being available by default to produce linting errors. To fix
this, it will be necessary to update the eslint config as described in the section "Linting Jest
code" in the README.